backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-multipart
passlib[bcrypt]
pyjwt